Filter splitting, more makefile fixes
