+++ .make.defaults
@@ -275,6 +275,7 @@ __check_defined = \
 .defaults.ray-lib-src-venv:: .defaults.create-venv .defaults.install-ray-lib-src-venv .defaults.install-local-requirements-venv
 
 # Install all source from the repo for a ray runtime transform into an existing venv 
+# And if there is an adjacent python dir (as for transforms), then also install that source
 .PHONY: .defaults.install-ray-lib-src-venv
 .defaults.install-ray-lib-src-venv:: 
 	@# Help: Install Ray and Python data processing library source into existing venv 
@@ -285,6 +286,9 @@ __check_defined = \
 	pip uninstall -y data-prep-toolkit-ray;		 		\
 	$(MAKE) PYTHON_PROJECT_DIR=$(DPK_PYTHON_LIB_DIR) .defaults.install-src-venv;		\
 	$(MAKE) PYTHON_PROJECT_DIR=$(DPK_RAY_LIB_DIR) .defaults.install-src-venv;		\
+	if [ -d ../python ]; then                                               \
+		$(MAKE) PYTHON_PROJECT_DIR=../python .defaults.install-src-venv;    \
+	fi 
 	echo Installed source from Ray data processing library for `which $(PYTHON)`
 
 # Install local requirements last as it generally includes our lib source
