+++ .make.versions
@@ -20,6 +20,8 @@ DOC_ID_VERSION=0.4.0$(RELEASE_VERSION_SUFFIX)
 DOC_ID_SPARK_VERSION=0.2.0$(RELEASE_VERSION_SUFFIX)
 EDEDUP_VERSION=0.4.0$(RELEASE_VERSION_SUFFIX)
 FDEDUP_VERSION=0.4.0$(RELEASE_VERSION_SUFFIX)
+FILTER_PYTHON_VERSION=0.4.0$(RELEASE_VERSION_SUFFIX)
+FILTER_RAY_VERSION=0.4.0$(RELEASE_VERSION_SUFFIX)
 FILTER_VERSION=0.4.0$(RELEASE_VERSION_SUFFIX)
 FILTER_SPARK_VERSION=0.2.0$(RELEASE_VERSION_SUFFIX)
 NOOP_PYTHON_VERSION=0.9.0$(RELEASE_VERSION_SUFFIX)
