+++ transforms/universal/noop/python/Dockerfile
@@ -26,6 +26,7 @@ RUN pip install --no-cache-dir -e .
 
 # copy source data
 COPY ./src/noop_transform.py .
+COPY ./src/noop_transform_python.py .
 COPY ./src/noop_local.py local/
 
 # copy test
